+++ backend/src/main/java/com/gptsam/core/common/exception/type/biz/BadRequestException.java
@@ -1,7 +1,7 @@
-package com.gptsam.core.exception.type.biz;
+package com.gptsam.core.common.exception.type.biz;
 
 
-import com.gptsam.core.exception.type.InternalServerException;
+import com.gptsam.core.common.exception.type.InternalServerException;
 
 /**
  * 400 Bad Request
