+++ backend/src/main/java/com/gptsam/core/common/exception/type/biz/ForbiddenException.java
@@ -1,6 +1,6 @@
-package com.gptsam.core.exception.type.biz;
+package com.gptsam.core.common.exception.type.biz;
 
-import com.gptsam.core.exception.type.InternalServerException;
+import com.gptsam.core.common.exception.type.InternalServerException;
 
 /**
  * 403 Forbidden
